refactor: core 모듈 정리 — enum을 :module:shared로 승격하고 lol-server-domain 제거
- module:core:enum → :module:shared 승격 (단일 자식만 감싸던 core 래퍼 제거)
- 패키지 com.example.lolserver → com.example.lolserver.shared (app 진입점과 split-package 해소)
- enum 5종 + TierFilter VO + 테스트 이동, import 21파일 / build.gradle 8개 / settings.gradle 갱신
- 빈 잔재 모듈 core:lol-server-domain 제거 (수직 컨텍스트 재구성 후 미사용, gitignore된 IDE 파일만 남아있던 상태)
- CLAUDE.md: 모듈 표 갱신 + 세션 학습(Docker 없는 검증 커맨드, settings.gradle 진실원천/stale 문서, Redis FQN 직렬화 gotcha)

검증: compileJava/compileTestJava 전 모듈, archTest(ArchUnit) 8모듈, :module:shared:test 8건, checkstyle 전 모듈 통과. (full test/build는 Docker 필요로 미실행)
