+++ module/domain/championstats/src/test/java/com/example/lolserver/championstats/application/ChampionStatsServiceConcurrencyTest.java
@@ -1,6 +1,6 @@
 package com.example.lolserver.championstats.application;
 
-import com.example.lolserver.TierFilter;
+import com.example.lolserver.shared.TierFilter;
 import com.example.lolserver.championstats.application.model.ChampionStatsReadModel;
 import com.example.lolserver.championstats.application.model.PositionChampionStatsReadModel;
 import com.example.lolserver.championstats.application.port.out.ChampionStatsCachePort;
