+++ module/domain/gamedata/build.gradle
@@ -5,7 +5,7 @@ plugins {
 
 dependencies {
     implementation project(':module:common')
-    implementation project(':module:core:enum')
+    implementation project(':module:shared')
     implementation project(':module:support:logging')
 
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
