+++ module/domain/gamedata/src/main/java/com/example/lolserver/gamedata/application/TierCutoffService.java
@@ -1,6 +1,6 @@
 package com.example.lolserver.gamedata.application;
 
-import com.example.lolserver.Platform;
+import com.example.lolserver.shared.Platform;
 import com.example.lolserver.gamedata.application.model.TierCutoffReadModel;
 import com.example.lolserver.gamedata.application.port.in.TierCutoffQueryUseCase;
 import com.example.lolserver.gamedata.application.port.out.TierCutoffPersistencePort;
