+++ module/domain/leaderboard/build.gradle
@@ -5,7 +5,7 @@ plugins {
 
 dependencies {
     implementation project(':module:common')
-    implementation project(':module:core:enum')
+    implementation project(':module:shared')
 
     testImplementation(testFixtures(project(':module:common')))
     testImplementation 'com.tngtech.archunit:archunit-junit5:1.3.0'
