+++ module/shared/src/main/java/com/example/lolserver/shared/QueueType.java
@@ -1,4 +1,4 @@
-package com.example.lolserver;
+package com.example.lolserver.shared;
 
 import lombok.Getter;
