+++ module/shared/src/main/java/com/example/lolserver/shared/RenewalStatus.java
@@ -1,4 +1,4 @@
-package com.example.lolserver;
+package com.example.lolserver.shared;
 
 public enum RenewalStatus {
     SUCCESS, FAILED, PROGRESS
