+++ module/shared/src/test/java/com/example/lolserver/shared/TierFilterTest.java
@@ -1,4 +1,4 @@
-package com.example.lolserver;
+package com.example.lolserver.shared;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
